typeurl: make marshal_test.go compile again and then pass
The test was importing the old import path instead of the new
import path, so it wasn't testing the code in this directory.
(And the code it was trying to test no longer exists.)

After making the test build, change the expected type
identifiers (per f694355) to make it also pass.