Remove gogo registry
Since containerd is not using gogo anymore it seems like we should
remove the depenency on gogo here since gogo is a fairly large
dependency and is also unaintained.

This also gets imported by the runc shim and adds a fair amount to the
binary size (in terms of percentage of the whole size).

I did keep a fallback test in that still imports gogo and continues to
work. The test marshals a type with gogo and then unmarshals it by
typeurl.
This is why gogo is still in the go.mod
Not sure if this is worthwhile or not.

This may need a module version bump since some functions will behave
differently now.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>