Fix unmarshalling of registered types
In 7d3d25864aa9e808fd5bb1a151ef70978d4fa675 I inadvterantly removed
support for json unmarshalling for the case when a type implements a
protobuf message but is a type that is registered.

For types that are registered through the `Register` function typeurl
is supposed to ignore the interfaces that the type implements and just
use json.

This change restores that behavior.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>