Fix unmarshalling of registered types
This is a faithful backport of #53

---

In 7d3d258 I inadvterantly removed
support for json unmarshalling for the case when a type implements a
protobuf message but is a type that is registered.

For types that are registered through the `Register` function typeurl
is supposed to ignore the interfaces that the type implements and just
use json.

This change restores that behavior.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>